rename imports, move a few more things around, fix type error in community
