+++ auth.ts
@@ -4,7 +4,7 @@ import NextAuth, { DefaultSession, type NextAuthConfig } from 'next-auth';
 import { encode } from 'next-auth/jwt';
 import microsoftEntraId from 'next-auth/providers/microsoft-entra-id';
 import { prisma } from '~/data/prisma';
-import { isDevelopment } from '~/lib/env';
+import { isDevelopment } from '~/data/env';
 import { UserService } from '~/features/users/user.service';
 
 declare module 'next-auth' {
