+++ src/features/users/user.service.ts
@@ -7,7 +7,7 @@ import { s3 } from '~/data/s3';
 import { withServiceAuth } from '~/auth/protected-service';
 import { ErrorMessage } from '~/lib/constants';
 import { AppError } from '~/lib/errors/app-error';
-import { isDevelopment } from '~/lib/env';
+import { isDevelopment } from '~/data/env';
 import {
   publicUserSelect,
   UpdateUserInput,
